Feature branch

+++ backend/foodgram/users/views.py
@@ -1,10 +1,9 @@
-from rest_framework import viewsets
-from rest_framework import status
-from rest_framework.response import Response
+from rest_framework import viewsets, status
 from rest_framework.decorators import action
+from rest_framework.response import Response
 from django.contrib.auth import update_session_auth_hash
 from django.shortcuts import get_object_or_404
-from rest_framework.permissions import IsAuthenticated
+from rest_framework.permissions import IsAuthenticated, AllowAny
 
 from users.models import User, Subscription
 from api.serializers import (
@@ -14,11 +13,14 @@ from api.serializers import (
     SubscriptionSerializer,
     SubscriptionActionSerializer
 )
+from recipes.paginations import CustomPaginator
+
 
 class CustomUserViewSet(viewsets.ModelViewSet):
     queryset = User.objects.all()
     serializer_class = CustomUserSerializer
-    permission_classes = [IsAuthenticated]
+    permission_classes = [AllowAny]
+    pagination_class = CustomPaginator
 
     def get_serializer_class(self):
         if self.action in ('list', 'retrieve'):
@@ -26,18 +28,28 @@ class CustomUserViewSet(viewsets.ModelViewSet):
         return CustomUserCreateSerializer
 
     def create(self, request, *args, **kwargs):
-        serializer = self.get_serializer(data=request.data, context={'request': request})
+        if 'first_name' not in request.data or 'last_name' not in request.data:
+            return Response(
+                {'error': 'Поля "first_name" и "last_name"'
+                 'обязательны для регистрации пользователя'},
+                status=status.HTTP_400_BAD_REQUEST)
+        serializer = self.get_serializer(data=request.data,
+                                         context={'request': request})
         serializer.is_valid(raise_exception=True)
         self.perform_create(serializer)
         headers = self.get_success_headers(serializer.data)
-        return Response(serializer.data, status=status.HTTP_201_CREATED, headers=headers)
+        return Response(
+            serializer.data, status=status.HTTP_201_CREATED, headers=headers)
 
-    @action(detail=False, methods=['get'], permission_classes=(IsAuthenticated,))
+    @action(detail=False, methods=['get'],
+            permission_classes=(IsAuthenticated,))
     def me(self, request):
-        serializer = CustomUserSerializer(request.user, context={'request': request})
+        serializer = CustomUserSerializer(request.user,
+                                          context={'request': request})
         return Response(serializer.data, status=status.HTTP_200_OK)
 
-    @action(detail=False, methods=['post'], permission_classes=(IsAuthenticated,))
+    @action(detail=False, methods=['post'],
+            permission_classes=(IsAuthenticated,))
     def set_password(self, request):
         serializer = CustomSetPasswordSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
@@ -45,20 +57,22 @@ class CustomUserViewSet(viewsets.ModelViewSet):
         current_password = serializer.validated_data['current_password']
         new_password = serializer.validated_data['new_password']
         if not user.check_password(current_password):
-            return Response({'error': 'Неверный текущий пароль'}, status=status.HTTP_400_BAD_REQUEST)
+            return Response({'error': 'Неверный текущий пароль'},
+                            status=status.HTTP_400_BAD_REQUEST)
         user.set_password(new_password)
         user.save()
         update_session_auth_hash(request, user)
-        return Response({'detail': 'Пароль успешно обновлен'}, status=status.HTTP_204_NO_CONTENT)
+        return Response({'detail': 'Пароль успешно обновлен'},
+                        status=status.HTTP_204_NO_CONTENT)
 
     @action(detail=False, methods=['get'],
-            permission_classes=(IsAuthenticated,),)
-            # pagination_class=CustomPaginator)
+            permission_classes=(IsAuthenticated,),
+            pagination_class=CustomPaginator)
     def subscriptions(self, request):
         queryset = User.objects.filter(subscription_author__user=request.user)
         page = self.paginate_queryset(queryset)
         serializer = SubscriptionSerializer(page, many=True,
-                                             context={'request': request})
+                                            context={'request': request})
         return self.get_paginated_response(serializer.data)
 
     @action(detail=True, methods=['post'],
@@ -68,20 +82,26 @@ class CustomUserViewSet(viewsets.ModelViewSet):
         if request.user == author:
             return Response({'error': 'Вы не можете подписаться на себя'},
                             status=status.HTTP_400_BAD_REQUEST)
-        subscription, created = Subscription.objects.get_or_create(user=request.user, author=author)
+        subscription, created = Subscription.objects.get_or_create(
+            user=request.user, author=author)
         if not created:
-            return Response({'error': 'Вы уже подписаны на этого пользователя'},
-                            status=status.HTTP_400_BAD_REQUEST)
-        serializer = SubscriptionActionSerializer(author, context={'request': request})
-        data = serializer.data
-        data['is_subscribed'] = True
-        return Response(data, status=status.HTTP_201_CREATED)
+            return Response(
+                {'error': 'Вы уже подписаны на этого пользователя'},
+                status=status.HTTP_400_BAD_REQUEST)
 
-    @action(detail=True, methods=['delete'], permission_classes=(IsAuthenticated,))
-    def unsubscribe(self, request, **kwargs):
+    @subscribe.mapping.delete
+    def delete_subscribe(self, request, **kwargs):
         author = get_object_or_404(User, id=kwargs['pk'])
-        subscription = get_object_or_404(Subscription, user=request.user, author=author)
+        subscription = Subscription.objects.filter(
+            user=request.user, author=author).first()
+        if not subscription:
+            return Response({'errors': 'Подписка не существует'},
+                            status=status.HTTP_400_BAD_REQUEST)
         subscription.delete()
-
-        return Response({'detail': 'Успешная отписка'},
-                        status=status.HTTP_204_NO_CONTENT)
+        serializer = SubscriptionActionSerializer(
+            author, context={'request': request})
+        data = serializer.data
+        data['is_subscribed'] = request.method == 'POST'
+        return Response(
+            data, status=status.HTTP_201_CREATED
+            if request.method == 'POST' else status.HTTP_204_NO_CONTENT)
